requirements.txt
Flask==2.1.2
Flask_Assets==2.0
numpy==1.21.5
pandas==1.4.2
mysql-connector-python
mysqlclient
gunicorn